move parseClusterGbk to antismash module
